Revert "Add basic Python API"
