+++ env.yml
@@ -1,7 +1,6 @@
 name: mafia-engine
 dependencies:
   - python=3.6
-  - pip=19.1.1
   - nose=1.3.7
   - flake8
   - sphinx=1.7.7
@@ -15,4 +14,3 @@ dependencies:
     - sphinxcontrib-napoleon==0.6.1
     - sphinxcontrib-websupport==1.1.0
     - fastapi[all]
-    - Deprecated
