+++ mafia/api/vanilla.py
