--- mafia/core/errors.py
@@ -1,6 +1,0 @@
-class MafiaError(Exception):
-    """Base class for :mod:`mafia` errors."""
-
-
-class AmbiguousName(UserWarning):
-    """Multiple objects found for this name."""
