+++ mafia/premade/ability/kill.py
