+++ mafia/premade/ability/vote.py
