+++ mafia/premade/template/vanilla.py
