+++ mafia/state/role.py
